backend/requirements.txt
# Core RAG components
langchain
langchain-openai
langchain-qdrant
langchain-community
qdrant-client
fastembed

# Required for fastembed (CPU-only ONNX backend)
onnxruntime

# Reranking
sentence-transformers

# API + Environment
fastapi
uvicorn[standard]
python-dotenv
tiktoken

# PDF & DOCX handling
pymupdf
pdfplumber
docx

# Evaluation tools
requests
nltk
rouge-score
bert-score[torch]

# Misc
pandas
psutil
wordcloud